+++ scripts/tools/note.js
@@ -858,7 +858,6 @@ function detectSlowSlider(diff, mapSettings) {
                 }
                 lastRed = note;
                 if (speedR > tool.minSliderSpeed) {
-                    console.log(lastRedTime);
                     arr.push(adjustTime(lastRedTime, bpm, offset, bpmc));
                 }
             } else {
@@ -876,7 +875,6 @@ function detectSlowSlider(diff, mapSettings) {
                 }
                 lastBlue = note;
                 if (speedB > tool.minSliderSpeed) {
-                    console.log(lastBlueTime, 'blue');
                     arr.push(adjustTime(lastBlueTime, bpm, offset, bpmc));
                 }
             } else {